+++ bridge/bindings/jsc/DOM/document.h
@@ -11,21 +11,21 @@
 #include "element.h"
 #include "node.h"
 
-
+#include "bindings/jsc/DOM/custom_event.h"
 #include "bindings/jsc/DOM/elements/anchor_element.h"
 #include "bindings/jsc/DOM/elements/canvas_element.h"
 #include "bindings/jsc/DOM/elements/image_element.h"
 #include "bindings/jsc/DOM/elements/input_element.h"
 #include "bindings/jsc/DOM/elements/object_element.h"
 #include "bindings/jsc/DOM/elements/script_element.h"
+#include "bindings/jsc/DOM/events/close_event.h"
+#include "bindings/jsc/DOM/events/gesture_event.h"
 #include "bindings/jsc/DOM/events/input_event.h"
+#include "bindings/jsc/DOM/events/intersection_change_event.h"
 #include "bindings/jsc/DOM/events/media_error_event.h"
 #include "bindings/jsc/DOM/events/message_event.h"
-#include "bindings/jsc/DOM/events/close_event.h"
-#include "bindings/jsc/DOM/events/intersection_change_event.h"
+#include "bindings/jsc/DOM/events/pop_state_event.h"
 #include "bindings/jsc/DOM/events/touch_event.h"
-#include "bindings/jsc/DOM/events/gesture_event.h"
-#include "bindings/jsc/DOM/custom_event.h"
 
 namespace kraken::binding::jsc {